+++ changelog.md
@@ -7,6 +7,7 @@
 - customizable toolbar items transitions (in/out).
 - customizable dock items transitions (in/out).
 - detailed information of WLAN when devtools enabled.
+- detailed information of bluetooth when devtools enabled.
 
 ### fix